Corrijo bugs de start y stop

+++ Stop.sh
@@ -6,18 +6,33 @@
 # >$ Stop.sh RecPro
 #------------------------------------------------------------------------------------------------------------
 
+source IniFunctions.sh
 
-PID=`ps aux | grep $BINDIR/$1.sh | grep -v grep | awk '{print $2}'`
+if [ ! -f $BINDIR/$1.sh ]; then
+    echo "No existe script $1"
+    exit 1
+fi
+
+#PID=`ps aux | grep "\($BINDIR\)\?/$1.sh" | grep -v grep | awk '{print $2}' | head -n 1`
+PID=$(getPid $1)
 
 if [ "$PID" != "" ]; then
-	kill -9 $PID
+    echo "Deteniendo proceso con pid $PID .."
+    kill -9 $PID
+    
+    if [ $? -ne 0 ];
+    then
+        echo "Error al detener el demonio con pid $PID"    
+    fi
 
-	PID=`ps aux | grep $BINDIR/$1.sh | grep -v grep | awk '{print $2}'`
+    #PID=`ps aux | grep "\($BINDIR\)\?/$1.sh" | grep -v grep | awk '{print $2}'`
+    #PID=`ps aux | grep "\($BINDIR\)\?/$1.sh" | grep -v grep | awk '{print $2}' | head -n 1`
+    PID=$(getPid $1)
 
-	if [ "$PID" != "" ]; then
-	    echo "No se pudo detener el demonio"    
-    	exit 1
-	fi
+    if [ "$PID" != "" ]; then
+        echo "No se pudo detener el demonio"    
+        exit 1
+    fi
 
     echo "Se detuvo el demonio"
     exit 0
